Refactor notification service tests to use prototype for dispatch spying and enhance event handling
