+++ app/backend/src/ingestion/__tests__/soroban-event.parser.unit.spec.ts
@@ -1,304 +1,168 @@
-import { EventEmitter2 } from "@nestjs/event-emitter";
-import { Test, TestingModule } from "@nestjs/testing";
-
-import { AppConfigService } from "../../config";
-import { CursorRepository } from "../cursor.repository";
-import { EscrowEventRepository } from "../escrow-event.repository";
+import { xdr, Address, nativeToScVal } from "stellar-sdk";
 import {
   SorobanEventParser,
   RawHorizonContractEvent,
 } from "../soroban-event.parser";
-import { StellarIngestionService } from "../stellar-ingestion.service";
-import type { EscrowDepositedEvent } from "../types/contract-event.types";
 
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
+function symVal(s: string): xdr.ScVal {
+  return xdr.ScVal.scvSymbol(s);
+}
 
-function makeRawEvent(
+function addressVal(pubkey: string): xdr.ScVal {
+  return Address.fromString(pubkey).toScVal();
+}
+
+function bytesVal(hex: string): xdr.ScVal {
+  return xdr.ScVal.scvBytes(Buffer.from(hex, "hex"));
+}
+
+function mapVal(entries: Record<string, xdr.ScVal>): xdr.ScVal {
+  const mapEntries = Object.entries(entries).map(
+    ([k, v]) => new xdr.ScMapEntry({ key: xdr.ScVal.scvSymbol(k), val: v }),
+  );
+  return xdr.ScVal.scvMap(mapEntries);
+}
+
+function makeRaw(
+  topics: xdr.ScVal[],
+  data: xdr.ScVal,
   overrides: Partial<RawHorizonContractEvent> = {},
 ): RawHorizonContractEvent {
   return {
-    id: "evt-1",
+    id: "1",
     paging_token: "100-1",
-    transaction_hash: "abc123",
+    transaction_hash: "txhash1",
     ledger: 100,
     created_at: "2025-01-01T00:00:00Z",
-    contract_id: "CTEST",
+    contract_id: "CAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF",
     type: "contract",
-    topic: [],
-    value: { xdr: "" },
+    topic: topics.map((v) => v.toXDR("base64")),
+    value: { xdr: data.toXDR("base64") },
     ...overrides,
   };
 }
 
-function makeEscrowDepositedEvent(
-  overrides: Partial<EscrowDepositedEvent> = {},
-): EscrowDepositedEvent {
-  return {
-    eventType: "EscrowDeposited",
-    txHash: "abc123",
-    ledgerSequence: 100,
-    pagingToken: "100-1",
-    contractTimestamp: 1700000000n,
-    commitment: "deadbeef".repeat(8),
-    owner: "GABC",
-    token: "CTOKEN",
-    amount: 1000000n,
-    expiresAt: 1800000000n,
-    ...overrides,
-  };
-}
+const OWNER = "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN";
+const TOKEN = "CDLZFC3SYJYDZT7K67VZ75HPJVIEUVNIXF47ZG2FB2RMQQVU2HHGCYSC";
+const COMMITMENT_HEX = "deadbeef".repeat(8);
 
-// ---------------------------------------------------------------------------
-// Mock factories
-// ---------------------------------------------------------------------------
+describe("SorobanEventParser", () => {
+  let parser: SorobanEventParser;
 
-const mockCursorRepo = (): jest.Mocked<CursorRepository> =>
-  ({
-    getCursor: jest.fn().mockResolvedValue(null),
-    saveCursor: jest.fn().mockResolvedValue(undefined),
-  }) as unknown as jest.Mocked<CursorRepository>;
-
-const mockEscrowRepo = (): jest.Mocked<EscrowEventRepository> =>
-  ({
-    upsertEvent: jest.fn().mockResolvedValue(undefined),
-  }) as unknown as jest.Mocked<EscrowEventRepository>;
-
-const mockParser = (): jest.Mocked<SorobanEventParser> =>
-  ({
-    parse: jest.fn().mockReturnValue(null),
-  }) as unknown as jest.Mocked<SorobanEventParser>;
-
-const mockConfig = (): Partial<AppConfigService> => ({
-  network: "testnet",
-});
-
-// ---------------------------------------------------------------------------
-// Tests
-// ---------------------------------------------------------------------------
-
-describe("StellarIngestionService", () => {
-  let service: StellarIngestionService;
-  let cursorRepo: jest.Mocked<CursorRepository>;
-  let escrowRepo: jest.Mocked<EscrowEventRepository>;
-  let parser: jest.Mocked<SorobanEventParser>;
-  let eventEmitter: EventEmitter2;
-
-  // Capture the stream callbacks installed by the service
-  let capturedOnMessage: ((record: RawHorizonContractEvent) => void) | null =
-    null;
-  let capturedOnError: ((err: unknown) => void) | null = null;
-  const mockStop = jest.fn();
-
-  beforeEach(async () => {
-    cursorRepo = mockCursorRepo();
-    escrowRepo = mockEscrowRepo();
-    parser = mockParser();
-    capturedOnMessage = null;
-    capturedOnError = null;
-    mockStop.mockClear();
-
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [
-        StellarIngestionService,
-        { provide: AppConfigService, useValue: mockConfig() },
-        { provide: CursorRepository, useValue: cursorRepo },
-        { provide: EscrowEventRepository, useValue: escrowRepo },
-        { provide: SorobanEventParser, useValue: parser },
-        EventEmitter2,
-      ],
-    }).compile();
-
-    service = module.get(StellarIngestionService);
-    eventEmitter = module.get(EventEmitter2);
-
-    // Stub the Horizon.Server used internally so we never touch the network
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    (service as any).server = {
-      contractEvents: jest.fn().mockImplementation(() => ({
-        cursor: jest.fn().mockReturnThis(),
-        stream: jest.fn().mockImplementation(({ onmessage, onerror }) => {
-          capturedOnMessage = onmessage as (
-            record: RawHorizonContractEvent,
-          ) => void;
-          capturedOnError = onerror as (err: unknown) => void;
-          return mockStop;
-        }),
-      })),
-    };
+  beforeEach(() => {
+    parser = new SorobanEventParser();
   });
 
-  afterEach(() => {
-    service.onModuleDestroy();
-  });
+  describe("EscrowDeposited", () => {
+    it("parses a valid EscrowDeposited event", () => {
+      const topics = [
+        symVal("EscrowDeposited"),
+        bytesVal(COMMITMENT_HEX),
+        addressVal(OWNER),
+      ];
+      const data = mapVal({
+        token: addressVal(TOKEN),
+        amount: nativeToScVal(5_000_000n, { type: "i128" }),
+        expires_at: nativeToScVal(1800000000n, { type: "u64" }),
+        timestamp: nativeToScVal(1700000000n, { type: "u64" }),
+      });
 
-  // -------------------------------------------------------------------------
-  // Cursor resumption
-  // -------------------------------------------------------------------------
-
-  describe("cursor resumption", () => {
-    it('starts from "now" when no cursor exists', async () => {
-      cursorRepo.getCursor.mockResolvedValue(null);
-      await service.startStreaming("CTEST");
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const server = (service as any).server as ReturnType<typeof jest.fn>;
-      expect(server.contractEvents).toHaveBeenCalledWith("CTEST");
-    });
-
-    it("resumes from stored cursor when one exists", async () => {
-      cursorRepo.getCursor.mockResolvedValue("99-5");
-      await service.startStreaming("CTEST");
-      expect(cursorRepo.getCursor).toHaveBeenCalledWith("contract:CTEST");
+      const result = parser.parse(makeRaw(topics, data));
+      expect(result?.eventType).toBe("EscrowDeposited");
+      if (result?.eventType !== "EscrowDeposited") return;
+      expect(result.commitment).toBe(COMMITMENT_HEX);
+      expect(result.owner).toBe(OWNER);
+      expect(result.amount).toBe(5_000_000n);
+      expect(result.expiresAt).toBe(1800000000n);
+      expect(result.contractTimestamp).toBe(1700000000n);
     });
   });
 
-  // -------------------------------------------------------------------------
-  // Event processing
-  // -------------------------------------------------------------------------
-
-  describe("event processing", () => {
-    beforeEach(async () => {
-      await service.startStreaming("CTEST");
-    });
-
-    it("advances cursor even for unrecognised events", async () => {
-      parser.parse.mockReturnValue(null);
-      const raw = makeRawEvent({ paging_token: "101-1", ledger: 101 });
-      await capturedOnMessage!(raw);
-
-      expect(cursorRepo.saveCursor).toHaveBeenCalledWith(
-        "contract:CTEST",
-        "101-1",
-        101,
-      );
-      expect(escrowRepo.upsertEvent).not.toHaveBeenCalled();
-    });
-
-    it("persists EscrowDeposited and advances cursor", async () => {
-      const event = makeEscrowDepositedEvent();
-      parser.parse.mockReturnValue(event);
-
-      const raw = makeRawEvent();
-      await capturedOnMessage!(raw);
-
-      expect(escrowRepo.upsertEvent).toHaveBeenCalledWith(event);
-      expect(cursorRepo.saveCursor).toHaveBeenCalledWith(
-        "contract:CTEST",
-        "100-1",
-        100,
-      );
-    });
-
-    it("emits a domain event via EventEmitter2", async () => {
-      const event = makeEscrowDepositedEvent();
-      parser.parse.mockReturnValue(event);
-
-      const listener = jest.fn();
-      eventEmitter.on("stellar.EscrowDeposited", listener);
-
-      await capturedOnMessage!(makeRawEvent());
-
-      expect(listener).toHaveBeenCalledWith(event);
-    });
-
-    it("does NOT throw when cursor save fails (non-fatal)", async () => {
-      parser.parse.mockReturnValue(null);
-      cursorRepo.saveCursor.mockRejectedValue(new Error("DB down"));
-
-      await expect(capturedOnMessage!(makeRawEvent())).resolves.not.toThrow();
+  describe("EscrowWithdrawn", () => {
+    it("parses a valid EscrowWithdrawn event", () => {
+      const topics = [
+        symVal("EscrowWithdrawn"),
+        bytesVal(COMMITMENT_HEX),
+        addressVal(OWNER),
+      ];
+      const data = mapVal({
+        token: addressVal(TOKEN),
+        amount: nativeToScVal(5_000_000n, { type: "i128" }),
+        timestamp: nativeToScVal(1700001000n, { type: "u64" }),
+      });
+      const result = parser.parse(makeRaw(topics, data));
+      expect(result?.eventType).toBe("EscrowWithdrawn");
     });
   });
 
-  // -------------------------------------------------------------------------
-  // Idempotency
-  // -------------------------------------------------------------------------
-
-  describe("idempotency", () => {
-    it("calls upsertEvent (ON CONFLICT DO NOTHING) on duplicate events", async () => {
-      const event = makeEscrowDepositedEvent();
-      parser.parse.mockReturnValue(event);
-      escrowRepo.upsertEvent.mockResolvedValue(undefined); // idempotent – no error on duplicate
-
-      await service.startStreaming("CTEST");
-
-      const raw = makeRawEvent();
-      await capturedOnMessage!(raw);
-      await capturedOnMessage!(raw); // replay same event
-
-      // Both calls must succeed; DB layer handles deduplication
-      expect(escrowRepo.upsertEvent).toHaveBeenCalledTimes(2);
+  describe("EscrowRefunded", () => {
+    it("parses a valid EscrowRefunded event", () => {
+      const topics = [
+        symVal("EscrowRefunded"),
+        bytesVal(COMMITMENT_HEX),
+        addressVal(OWNER),
+      ];
+      const data = mapVal({
+        token: addressVal(TOKEN),
+        amount: nativeToScVal(5_000_000n, { type: "i128" }),
+        timestamp: nativeToScVal(1700002000n, { type: "u64" }),
+      });
+      const result = parser.parse(makeRaw(topics, data));
+      expect(result?.eventType).toBe("EscrowRefunded");
     });
   });
 
-  // -------------------------------------------------------------------------
-  // Stream disconnection and reconnection
-  // -------------------------------------------------------------------------
+  describe("PrivacyToggled", () => {
+    it("parses a valid PrivacyToggled event", () => {
+      const topics = [symVal("PrivacyToggled"), addressVal(OWNER)];
+      const data = mapVal({
+        enabled: nativeToScVal(true),
+        timestamp: nativeToScVal(1700003000n, { type: "u64" }),
+      });
+      const result = parser.parse(makeRaw(topics, data));
+      expect(result?.eventType).toBe("PrivacyToggled");
+      if (result?.eventType !== "PrivacyToggled") return;
+      expect(result.enabled).toBe(true);
+      expect(result.owner).toBe(OWNER);
+    });
+  });
 
-  describe("reconnection", () => {
-    jest.useFakeTimers();
+  describe("AdminChanged", () => {
+    it("parses a valid AdminChanged event", () => {
+      const ADMIN2 = "GBVVJJWOR35BPXM2XJQLMQBDNKJWKJNPGQLDNPVKPOUJDMKBDLKMNKR";
+      const topics = [
+        symVal("AdminChanged"),
+        addressVal(OWNER),
+        addressVal(ADMIN2),
+      ];
+      const data = mapVal({
+        timestamp: nativeToScVal(1700004000n, { type: "u64" }),
+      });
+      const result = parser.parse(makeRaw(topics, data));
+      expect(result?.eventType).toBe("AdminChanged");
+      if (result?.eventType !== "AdminChanged") return;
+      expect(result.oldAdmin).toBe(OWNER);
+      expect(result.newAdmin).toBe(ADMIN2);
+    });
+  });
 
-    it("schedules a reconnect when stream emits an error", async () => {
-      await service.startStreaming("CTEST");
-      const openStreamSpy = jest.spyOn(
-        service as unknown as { openStream: () => Promise<void> },
-        "openStream" as never,
-      );
-
-      capturedOnError!(new Error("Connection reset"));
-
-      // Back-off timer should be scheduled
-      jest.advanceTimersByTime(1_100);
-      // openStream would be called again
-      expect(openStreamSpy).toHaveBeenCalledTimes(1);
+  describe("error cases", () => {
+    it("returns null for an event with no topics", () => {
+      expect(parser.parse(makeRaw([], xdr.ScVal.scvVoid()))).toBeNull();
     });
 
-    it("uses exponential back-off on repeated failures", async () => {
-      await service.startStreaming("CTEST");
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const svc = service as any;
-
-      // Simulate multiple disconnects
-      capturedOnError!(new Error("disconnect 1"));
-      expect(svc.currentBackoffMs).toBe(2_000); // doubled from 1_000
-
-      // Re-open stream
-      jest.advanceTimersByTime(2_100);
-      capturedOnError!(new Error("disconnect 2"));
-      expect(svc.currentBackoffMs).toBe(4_000);
-
-      capturedOnError!(new Error("disconnect 3"));
-      expect(svc.currentBackoffMs).toBe(8_000);
+    it("returns null for an unrecognised event name", () => {
+      expect(
+        parser.parse(makeRaw([symVal("UnknownEvent")], xdr.ScVal.scvVoid())),
+      ).toBeNull();
     });
 
-    it("caps backoff at MAX_BACKOFF_MS (60 s)", async () => {
-      await service.startStreaming("CTEST");
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const svc = service as any;
-      svc.currentBackoffMs = 32_000;
-
-      capturedOnError!(new Error("disconnect"));
-      expect(svc.currentBackoffMs).toBe(60_000); // capped
-    });
-
-    it("does not reconnect after onModuleDestroy()", async () => {
-      await service.startStreaming("CTEST");
-      service.onModuleDestroy();
-
-      const openStreamSpy = jest.spyOn(
-        service as unknown as { openStream: () => Promise<void> },
-        "openStream" as never,
-      );
-      capturedOnError!(new Error("late error"));
-
-      jest.advanceTimersByTime(2_000);
-      expect(openStreamSpy).not.toHaveBeenCalled();
-    });
-
-    afterEach(() => {
-      jest.useRealTimers();
+    it("returns null and does not throw on malformed XDR", () => {
+      const raw = makeRaw([], xdr.ScVal.scvVoid(), {
+        topic: ["not-valid-base64!!!"],
+      });
+      expect(() => parser.parse(raw)).not.toThrow();
+      expect(parser.parse(raw)).toBeNull();
     });
   });
 });
